+++ tests/testthat/test-utils.R
@@ -28,8 +28,7 @@ test_that("utils", {
   
   expect_identical(npars(boron_lnorm), 2L)
   expect_identical(npars(boron_dists), c(
-    gamma = 2L, gompertz = 2L, lgumbel = 2L,
-    llog = 2L, lnorm = 2L, weibull = 2L
+    burrIII2 = 2L, gamma = 2L, lnorm = 2L
   ))
   expect_identical(npars(fluazinam_lnorm), 2L)
   expect_identical(