+++ NEWS.md
@@ -1,3 +1,5 @@
+- Default distributions changed to 'burrIII2', 'gamma' and 'lnorm' from
+'gamma', 'gompertz', 'lgumbel', 'llog', 'lnorm' and 'weibull'.
 - `ssd_fit_dists()` now checks if standard errors computable.
 - Added `sdist(x)` functionality to set starting values for distributions.
 - Added Burr Type-III Three-Parameter Distribution (`burrIII3`).